\name{explain.harmonicMean}

\alias{explain.harmonicMean}

\title{Harmonic Mean Function Explained}

\description{Step by step demonstration of the harmonic mean calculus.}

\usage{
explain.harmonicMean(x)
}

\arguments{
  \item{x}{Should be a numbers vector}
}

\details{To calculate the harmonic mean, the user should give a numbers vector. The result is the explained process to calculate the harmonic mean, with the data of the dataset provided like argument. We can saw the harmonic mean formule in the harmonicMean_ help document.}

\value{Numeric result and the process of this calculus explained.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

\note{A vector is created by c(), like c(1,2,3,4,5) creates a vector with the numbers: 1,2,3,4,5 }

%\seealso{}

\examples{
  
  #data creation
  data <- c(10,4,5,7,3,4,1)
  
  explain.harmonicMean(data)
    
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~harmonic }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~armonica }% __ONLY ONE__ keyword per line
\keyword{ ~explain }
\keyword{ ~mean }
\keyword{ ~media }
\keyword{ ~explicada}
