\name{explain.normal}

\alias{explain.normal}

\title{Normal Distribution Function Explained}

\description{Step by step demonstration of the normal distribution calculus.}

\usage{
explain.normal(x)
}

\arguments{
  \item{x}{Should be a number.}
}

\details{To calculate the normal distribution, the user should give a number. The result isis a type of continuous probability distribution for a real-valued random variable. We can saw the normal distribution correlation formule in the normal_ help document.}

\value{Numeric result and the process of this calculus explained.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

\note{The variable is a number. Example: x <- 0.1 }

%\seealso{}

\examples{

    #data creation
    x = 0.1
    
    explain.normal(x)
  
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~normal }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution }% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~explain }
\keyword{ ~explicada}
