\name{getUserAction}

\alias{getUserAction}

\title{Get User Action Funcion}

\description{This function get the buffer introduced by the user. Typically a numerical vector.}

\usage{
getUserAction()
}

%\details{}

\value{A vector}

%\references{}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

\note{The process is interactive with the user}

%\seealso{}

\examples{
{
  \dontrun{
  vector <- getUserAction()
  }
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~get }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~user }% __ONLY ONE__ keyword per line
\keyword{ ~coger}
\keyword{ ~usuario}
