\name{initImages}

\alias{initImages}

\title{Init Images Function}

\description{This function is used to display an image.}

\usage{
initImages(path)
}

\arguments{
  \item{path}{An url of an image}
}

%\details{}

\value{There isn't return value}

%\references{}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

\note{The path shold be toward an image}

%\seealso{}

\examples{
{

  \dontrun{
    path = "https://i.imgur.com/8237YhzJ.png"
  initImages(path)

  }
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~init }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~images }% __ONLY ONE__ keyword per line
\keyword{ ~iniciar}
\keyword{ ~imagenes}
