\name{interactive.absolute_acum_frecuency}

\alias{interactive.absolute_acum_frecuency}

\title{User Interactive Absolute Accumulated Frecuency Calculus}

\description{Interactive function for absolute accumulated frecuency calculus.}

\usage{
interactive.absolute_acum_frecuency()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the formula.}

\value{An interactive process to calculate the absolute accumulated frecuency}


\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{
    
  \dontrun{
  interactive.absolute_acum_frecuency()
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~absolute }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~frecuency}
\keyword{ ~accumulated}
