\name{interactive.averageDeviation}

\alias{interactive.averageDeviation}

\title{User Interactive Average Absolute Deviation Calculus}

\description{Interactive function for average absolute deviation calculus.}

\usage{
interactive.averageDeviation()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the average absolute deviation formule, apart from the averageDeviation_ help document.}

\value{An interactive process to calculate the average absolute deviation}

\references{\url{https://en.wikipedia.org/wiki/Average_absolute_deviation}}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
}

%\note{}
%\seealso{}

\examples{

   \dontrun{
  interactive.averageDeviation()
  }

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~deviation }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~average }% __ONLY ONE__ keyword per line
\keyword{ ~absolute}
\keyword{ ~interactive}
\keyword{ ~interactiva}
\keyword{ ~absoluta}
\keyword{ ~desviacion}