\name{interactive.binomial}

\alias{interactive.binomial}

\title{User Interactive Binomial Distribution Calculus}

\description{Interactive function for binomial distribution calculus.}

\usage{
interactive.binomial()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the values when the function needs it. After that, the function will ask what is the correct result for this datas. The function itself will provide the binomial distribution formule, apart from the binomial_ help document.}

\value{An interactive process to calculate the binomial distribution.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.binomial()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~binomial }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution}% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~interactive }
\keyword{ ~interactiva }
