\name{interactive.chisquared}

\alias{interactive.chisquared}

\title{User Interactive Chisquared Distribution Calculus}

\description{Interactive function for chisquared distribution calculus.}

\usage{
interactive.chisquared()
}

\details{The user provides the datasets when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the chisquared distribution formule, apart from the chisquared_ help document.}

\value{An interactive process to calculate the chisquared distribution}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.chisquared()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~chisquared }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~chicuadrado }% __ONLY ONE__ keyword per line
\keyword{ ~interactive }
\keyword{ ~interactiva }
