\name{interactive.covariance}

\alias{interactive.covariance}

\title{User Interactive Covariance Calculus}

\description{Interactive function for covariance calculus.}

\usage{
interactive.covariance()
}

\details{The user provides the datasets when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the covariance formule, apart from the covariance_ help document.}

\value{An interactive process to calculate the covariance}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.covariance()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~covariance }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~covarianza }% __ONLY ONE__ keyword per line
\keyword{ ~interactive }
\keyword{ ~interactiva }
