\name{interactive.cv}

\alias{interactive.cv}

\title{User Interactive Coefficient of Variation Calculus}

\description{Interactive function for Coefficient of Variation calculus.}

\usage{
interactive.cv()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the coefficient of variation formule, apart from the cv_ help document.}

\value{An interactive process to calculate the average absolute deviation}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.cv()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~coefficient }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~variation }% __ONLY ONE__ keyword per line
\keyword{ ~coeficiente }
\keyword{ ~variacion }
\keyword{ ~cv }
\keyword{ ~interactive }
\keyword{ ~interactiva }
