\name{interactive.fisher}

\alias{interactive.fisher}

\title{User Interactive F Fisher Distribution Calculus}

\description{Interactive function for fisher distribution calculus.}

\usage{
interactive.fisher()
}

\details{The user provides the datasets when the function needs it. After that, the function will ask what is the correct result for this datasets. The function itself will provide the fisher distribution formule, apart from the fisher_ help document.}

\value{An interactive process to calculate the fisher distribution}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.fisher()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~fisher }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution }% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~interactive }
\keyword{ ~interactiva }
