\name{interactive.geometricMean}

\alias{interactive.geometricMean}

\title{User Interactive Geometric Mean Calculus}

\description{Interactive function for geometric mean calculus.}

\usage{
interactive.geometricMean()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the geometric mean formule, apart from the geometricMean_ help document.}

\value{An interactive process to calculate the geometric mean.}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.geometricMean()
  }
  
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~interactive }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~geometric }% __ONLY ONE__ keyword per line
\keyword{ ~mean}
\keyword{ ~interactivo}
\keyword{ ~geometrica}
\keyword{ ~media}
