\name{interactive.harmonicMean}

\alias{interactive.harmonicMean}

\title{User Interactive Harmonic Mean Calculus}

\description{Interactive function for harmonic mean calculus.}

\usage{
interactive.harmonicMean()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the harmonic mean formule, apart from the harmonicMean_ help document.}

\value{An interactive process to calculate the harmonic mean}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.harmonicMean()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~harmonic }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~armonica }% __ONLY ONE__ keyword per line
\keyword{ ~interactive }
\keyword{ ~mean }
\keyword{ ~media }
\keyword{ ~interactiva}
