\name{interactive.laplace}

\alias{interactive.laplace}

\title{User Interactive Laplace's Rule Calculus}

\description{Interactive function for Laplace's rule calculus.}

\usage{
interactive.laplace()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the values when the function needs it. After that, the function will ask what is the correct result for this datas. The function itself will provide the Laplace's rule formule, apart from the laplace_ help document.}

\value{An interactive process to calculate the Laplace's rule.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.laplace()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~laplace }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~rule }% __ONLY ONE__ keyword per line
\keyword{ ~regla }
\keyword{ ~interactive }
\keyword{ ~interactiva }
