\name{interactive.mean}

\alias{interactive.mean}

\title{User Interactive Mean Calculus}

\description{Interactive function for arithmetic mean calculus.}

\usage{
interactive.mean()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the arithmetic mean formule, apart from the mean_ help document.}

\value{An interactive process to calculate the arithmetic mean.}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{
{
  
  \dontrun{
  interactive.mean()
  }
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~mean }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~interactivo}
\keyword{ ~media.interactiva}