\name{interactive.median}

\alias{interactive.median}

\title{User Interactive Median Calculus}

\description{Interactive function for median calculus.}

\usage{
interactive.median()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the median formule, apart from the median_ help document.}

\value{An interactive process to calculate the median}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.median()
  }

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~median }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~interactiva}
\keyword{ ~mediana.interactiva}
