\name{interactive.mode}

\alias{interactive.mode}

\title{User Interactive Mode Calculus}

\description{Interactive function for mode calculus.}

\usage{
interactive.mode()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset.}

\value{An interactive process to calculate the mode.}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.mode()
  }

}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~mode }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~interactiva}
\keyword{ ~moda.interactiva}
