\name{interactive.normal}

\alias{interactive.normal}

\title{User Interactive Normal Distribution Calculus}

\description{Interactive function for normal distribution calculus.}

\usage{
interactive.normal()
}

\details{The user provides the values when the function needs it. After that, the function will ask what is the correct result for this data. The function itself will provide the normal distribution formule, apart from the normal_ help document.}

\value{An interactive process to calculate the normal distribution.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.normal()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~normal }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution }% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~interactive }
\keyword{ ~interactiva }
