\name{interactive.pearson}

\alias{interactive.pearson}

\title{User Interactive Pearson Correlation Calculus}

\description{Interactive function for pearson correlation calculus.}

\usage{
interactive.pearson()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the datasets when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the pearson correlation formule, apart from the pearson_ help document.}

\value{An interactive process to calculate the pearson correlation.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.pearson()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~covariance }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~covarianza }% __ONLY ONE__ keyword per line
\keyword{ ~interactive }
\keyword{ ~interactiva }
