\name{interactive.percentile}

\alias{interactive.percentile}

\title{User Interactive Percentile Calculus}

\description{Interactive function for percentiles calculus.}

\usage{
interactive.percentile()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the formula.} 

\value{An interactive process to calculate the percentiles}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.percentile()
  }

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~percentil }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~percentile}
\keyword{ ~percentiles}
