\name{interactive.poisson}

\alias{interactive.poisson}

\title{User Interactive Poisson Distribution Calculus}

\description{Interactive function for Poisson distribution calculus.}

\usage{
interactive.poisson()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the values when the function needs it. After that, the function will ask what is the correct result for this datas. The function itself will provide the Poisson distribution formule, apart from the poisson_ help document.}

\value{An interactive process to calculate the Poisson distribution.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.poisson()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~poisson }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution }% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~interactive }
\keyword{ ~interactiva }
