\name{interactive.quartile}

\alias{interactive.quartile}

\title{User Interactive Quartiles Calculus}

\description{Interactive function for quartiles calculus.}

\usage{
interactive.quartile()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the formula.}

\value{An interactive process to calculate the quartiles}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.quartile()
  }

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~cuartiles }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~quartile}
\keyword{ ~quantiles}
