\name{interactive.relative_frecuency}

\alias{interactive.relative_frecuency}

\title{User Interactive Relative Frecuency Calculus}

\description{Interactive function for relative frecuency calculus.}

\usage{
interactive.relative_frecuency()
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the formula.}

\value{An interactive process to calculate the relative frecuency}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.relative_frecuency()
  }

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~relative }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~interactive }% __ONLY ONE__ keyword per line
\keyword{ ~frecuency}
\keyword{ ~frecuencia}
