\name{interactive.standardDeviation}

\alias{interactive.standardDeviation}

\title{User Interactive Standard Deviation Calculus}

\description{Interactive function for standard deviation calculus.}

\usage{
interactive.standardDeviation()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the dataset when the function needs it. After that, the function will ask what is the correct result for this dataset. The function itself will provide the standard deviation formule, apart from the standardDeviation_ help document.}

\value{An interactive process to calculate the standard deviation}

\references{\url{https://en.wikipedia.org/wiki/Standard_deviation}}

\author{
Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
}

%\note{}
%\seealso{}

\examples{

   \dontrun{
  interactive.standardDeviation()
  }
  
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~deviation }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~standard }% __ONLY ONE__ keyword per line
\keyword{ ~interactive}
\keyword{ ~interactiva}
\keyword{ ~tipica}
\keyword{ ~desviacion}
