\name{interactive.tstudent}

\alias{interactive.tstudent}

\title{User Interactive T-Student Distribution Calculus}

\description{Interactive function for T-Student distribution calculus.}

\usage{
interactive.tstudent()
}

\arguments{
  \item{}{void, not argument}
}

\details{The user provides the values when the function needs it. After that, the function will ask what is the correct result for this datas. The function itself will provide the T-Students distribution formule, apart from the tstudent_ help document.}

\value{An interactive process to calculate the T-Student distribution.}

\author{Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
}

%\note{}
%\seealso{}

\examples{

  \dontrun{
  interactive.tstudent()
  }
 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~tstudent }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~distribution }% __ONLY ONE__ keyword per line
\keyword{ ~distribucion }
\keyword{ ~interactive }
\keyword{ ~interactiva }
