\name{LearningRlab}
\alias{LearningRlab}
\docType{package}
\title{Statistical Learning Functions}
\description{Package used to teach basic statistics to students.}
\details{
    This package pretends to serve the user as a method of learning basic statistical functions at secondary and baccalaureate courses. The content of the package incorporate a serie of statistical functions like the calculus of the arithmetic mean or the calculus of the frequencies.
There is no only calculus functions, further more, there are incorporated interactive and explicative functions to help and guide the user in the learning process.
  \tabular{ll}{
    Package: \tab LearningRlab\cr
    Type: \tab Package\cr
    Version: \tab 2.2\cr
    Date: \tab 2021-1-17\cr
    License: \tab Unlimited\cr
  }
}
\author{
Jose Manuel Gomez Caceres, \email{josemanuel.gomezc@edu.uah.es}
\cr{Dennis Monheimius, \email{dennis.monhemimius@edu.uah.es}}
\cr{Eduardo Benito, \email{eduardo.benito@edu.uah.es}}
\cr{Juan Jose Cuadrado, \email{jjcg@uah.es}}
\cr{Universidad de Alcala de Henares}
\cr {Maintainer: Eduardo Benito \email{eduardo.benito@edu.uah.com}}
}
